Remove the user_ref['password'] element if LDAP returns it
The logic for determining if a user was found in SQL is to check
to see if the key password is in the returned user_ref dict:

except KeyError:  # if it doesn't have a password, it must be LDAP

If, however, the LDAP server returns a password element, this
except will fail, causing us to attempt to validate an LDAP
user against SQL. This will almost always fail for obvious reasons.

Removing the password element if it exists in the dict returned by
LDAP alleviates this issue.